refactor: split nix specific code into seperate module
